+++ packages/libs/lambda-at-edge/src/default-handler.ts
@@ -24,7 +24,12 @@ import { performance } from "perf_hooks";
 import { ServerResponse } from "http";
 import jsonwebtoken from "jsonwebtoken";
 import type { Readable } from "stream";
-import { createRedirectResponse, getRedirectPath } from "./routing/redirector";
+import {
+  createRedirectResponse,
+  getDomainRedirectPath,
+  getRedirectPath
+} from "./routing/redirector";
+import { getRewritePath } from "./routing/rewriter";
 
 const basePath = RoutesManifestJson.basePath;
 const NEXT_PREVIEW_DATA_COOKIE = "__next_preview_data";
@@ -208,9 +213,16 @@ const handleOriginRequest = async ({
   prerenderManifest: PrerenderManifestType;
   routesManifest: RoutesManifest;
 }) => {
-  const basePath = routesManifest.basePath;
   const request = event.Records[0].cf.request;
-  const uri = normaliseUri(request.uri);
+
+  // Handle domain redirects e.g www to non-www domain
+  const domainRedirect = getDomainRedirectPath(request, manifest);
+  if (domainRedirect) {
+    return createRedirectResponse(domainRedirect, request.querystring, 308);
+  }
+
+  const basePath = routesManifest.basePath;
+  let uri = normaliseUri(request.uri);
   const { pages, publicFiles } = manifest;
   const isPublicFile = publicFiles[uri];
   const isDataReq = isDataRequest(uri);
@@ -253,6 +265,13 @@ const handleOriginRequest = async ({
       request.querystring,
       customRedirect.statusCode
     );
+  }
+
+  // Handle custom rewrites
+  const customRewrite = getRewritePath(request.uri, routesManifest);
+  if (customRewrite) {
+    request.uri = customRewrite;
+    uri = normaliseUri(request.uri);
   }
 
   const isStaticPage = pages.html.nonDynamic[uri];