+++ packages/serverless-components/nextjs-component/types.d.ts
@@ -18,6 +18,7 @@ export type ServerlessComponentInputs = {
   policy?: string;
   domain?: string | string[];
   domainType?: "www" | "apex" | "both";
+  domainRedirects?: { [key: string]: string };
   cloudfront?: CloudfrontOptions;
 };